api/sdc/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
python-dotenv==1.0.0
aiofiles==23.2.1
aiofiles
alembic
fastapi
pandas
psycopg2-binary
pydantic
python-dotenv
python-multipart
sqlalchemy
uvicorn[standard]
pydantic-settings
django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
openpyxl==3.1.2
openai==1.107.0
aiohttp==3.9.1
azure-ai-formrecognizer==3.3.2
azure-ai-textanalytics==5.3.0
djangorestframework-csv
drf-excel
aiofiles
aiohttp
alembic
azure-ai-formrecognizer
azure-ai-textanalytics
django
django-cors-headers
djangorestframework
djangorestframework-csv
drf-excel
fastapi
openai
openpyxl
pandas
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-multipart
sqlalchemy
uvicorn[standard]
scikit-learn 
numpy
# AI/ML dependencies for Potential Fitments
scikit-learn>=1.0
numpy>=1.21
dj-database-url
django
django-cors-headers
djangorestframework
psycopg2-binary
python-dotenv
whitenoise
azure-storage-blob
chardet
django_filters
celery
redis